+++ magda-web-client/src/Components/Home/MyDatasetSectionComponents/DatasetList.tsx
@@ -12,6 +12,7 @@ import openWindow from "helpers/openWindow";
 import { useHistory, Link, useLocation } from "react-router-dom";
 import { Location } from "history";
 import urijs from "urijs";
+import { getUrlWithPopUpQueryString } from "helpers/popupUtils";
 
 type PropsType = {
     openInPopUp?: boolean;
@@ -76,7 +77,9 @@ const DatasetList: FunctionComponent<PropsType> = (props) => {
                             className={`${
                                 activeTab === "drafts" ? "active" : ""
                             }`}
-                            to="/settings/datasets/draft"
+                            to={getUrlWithPopUpQueryString(
+                                "/settings/datasets/draft"
+                            )}
                         >
                             Drafts
                         </Link>
@@ -85,7 +88,9 @@ const DatasetList: FunctionComponent<PropsType> = (props) => {
                             className={`${
                                 activeTab === "published" ? "active" : ""
                             }`}
-                            to="/settings/datasets/published"
+                            to={getUrlWithPopUpQueryString(
+                                "/settings/datasets/published"
+                            )}
                         >
                             Published
                         </Link>